+++ src/input/input.hpp
@@ -1,7 +1,11 @@
 #pragma once
 
+#include <orge.h>
+
 namespace input {
 
 void update();
+
+int32_t getState(OrgeScancode scancode);
 
 } // namespace input
